When played melody is asked - change meter on the score
- there is no guessing meter yet (if any)
- setting meter on the score determines visible metronome counting
  (if not set, counting is misleading)
- TscoreObject - cast meter enum to int

+++ src/libs/core/score/tscoreobject.cpp
@@ -497,7 +497,7 @@ void TscoreObject::setMelody(Tmelody* melody, bool ignoreTechnical, int notesAmo
 CHECKTIME (
   clearScorePrivate();
   m_notes.clear();
-  setMeter(melody->meter()->meter());
+  setMeter(static_cast<int>(melody->meter()->meter()));
   setClefType(melody->clef());
   int newKey = static_cast<int>(melody->key().value());
   if (newKey != keySignature()) {
